Properly escape path before using in regex
Currently, the pattern used to ensure `__init__.py`s are installed for
package subdirectories (e.g. msg, srv) uses an un-escaped path. As a
result, any special characters in any part of the path can throw off the
regex and cause these files to not be properly installed.

Before using the path in the pattern, properly escape those special
characters.

Signed-off-by: Kyle Fazzari <kyle@canonical.com>