pkg-genmsg.cmake: fix location of pycache obj
Without the -d switch, the host location of the source files
that get compiled with -m compileall gets injected into the object
files. This is unwanted behavior, as it leaks information of the
host and also does not provide correct information to the python
interpreter, as such files are no longer available on the host.

Signed-off-by: Matthias Schoepfer <m.schoepfer@rethinkrobotics.com>